@@ -1,17 +1,17 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <AssemblyName>Spear.Sharp</AssemblyName>
     <RootNamespace>Spear.Sharp</RootNamespace>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>bin\Debug\netcoreapp2.1\Acb.Spear.xml</DocumentationFile>
+    <DocumentationFile>bin\Debug\netcoreapp3.1\Acb.Spear.xml</DocumentationFile>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile>bin\Release\netcoreapp2.1\Spear.Sharp.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\netcoreapp3.1\Spear.Sharp.xml</DocumentationFile>
   </PropertyGroup>
 
   <ItemGroup>
@@ -40,11 +40,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Acb.Core" Version="1.6.7" />
-    <PackageReference Include="Acb.WebApi" Version="1.6.6" />
+    <PackageReference Include="Acb.Core" Version="1.8.4" />
+    <PackageReference Include="Acb.WebApi" Version="1.8.4" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR" Version="1.1.0" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Extensions.Http" Version="2.2.0" />
   </ItemGroup>
 
   <ItemGroup>
@@ -60,6 +58,9 @@
 
   <ItemGroup>
     <None Update="Dockerfile">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="push.sh">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
