@@ -23,20 +23,24 @@
                 </div>
                 <div class="col-sm-3 b-naming">
                     <select class="form-control">
-                        <option value="0">Naming</option>
-                        <option value="1">Table</option>
+                        <option value="0">None</option>
+                        <option value="1" selected="selected">Naming</option>
+                        <option value="2">Table</option>
+
                     </select>
                 </div>
             </div>
-            <div class="d-naming">
-                <pre>
-<code class="cs">@Model.EntityCode(provider, language, true)</code>
-                </pre>
+            <div class="code d-none hide">
+                <button class="btn btn-primary b-copy">Copy</button>
+                <pre><code class="cs">@Model.EntityCode(provider, language, NamingType.None)</code></pre>
             </div>
-            <div class="d-table hide">
-                <pre>
-<code class="cs">@Model.EntityCode(provider, language, false)</code>
-                </pre>
+            <div class="code d-naming">
+                <button class="btn btn-primary b-copy">Copy</button>
+                <pre><code class="cs">@Model.EntityCode(provider)</code></pre>
+            </div>
+            <div class="code d-table hide">
+                <button class="btn btn-primary b-copy">Copy</button>
+                <pre><code class="cs">@Model.EntityCode(provider, language, NamingType.Table)</code></pre>
             </div>
         </div>
     </div>
